requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
marshmallow==3.13.0
psycopg2-binary==2.9.1
pymysql==1.0.2
SQLAlchemy==1.4.23
pytest==6.2.5
python-dotenv==0.19.0
Werkzeug==2.0.1